Revert "refactors & more"
This reverts commit f4b6be24907972f0d9b8e4d5227e698a812e0dcd.
