+++ docker_images/sklearn/tests/generators/samples/iris-latest-input.json
@@ -1,2 +1,52 @@
 {
-  "data": 
+  "data": {
+    "sepal length (cm)": [
+      6.1,
+      5.7,
+      7.7,
+      6.0,
+      6.8,
+      5.4,
+      5.6,
+      6.9,
+      6.2,
+      5.8
+    ],
+    "sepal width (cm)": [
+      2.8,
+      3.8,
+      2.6,
+      2.9,
+      2.8,
+      3.4,
+      2.9,
+      3.1,
+      2.2,
+      2.7
+    ],
+    "petal length (cm)": [
+      4.7,
+      1.7,
+      6.9,
+      4.5,
+      4.8,
+      1.5,
+      3.6,
+      5.1,
+      4.5,
+      3.9
+    ],
+    "petal width (cm)": [
+      1.2,
+      0.3,
+      2.3,
+      1.5,
+      1.4,
+      0.4,
+      1.3,
+      2.3,
+      1.5,
+      1.2
+    ]
+  }
+}
