+++ docker_images/sklearn/tests/generators/samples/iris-logistic_regression-1.0.0-output.json
@@ -1,0 +1,12 @@
+[
+  1,
+  0,
+  2,
+  1,
+  1,
+  0,
+  1,
+  2,
+  1,
+  1
+]
