+++ docker_images/sklearn/tests/generators/samples/tabularregression-1.0.0-input.json
@@ -1,0 +1,124 @@
+{
+  "data": {
+    "age": [
+      0.0453409833354632,
+      0.0925639831987174,
+      0.063503675590561,
+      0.096196521649737,
+      0.0126481372762872,
+      0.00901559882526763,
+      -0.00914709342983014,
+      -0.0236772472339084,
+      -0.0926954778032799,
+      -0.0600026317441039
+    ],
+    "sex": [
+      -0.044641636506989,
+      -0.044641636506989,
+      0.0506801187398187,
+      -0.044641636506989,
+      0.0506801187398187,
+      -0.044641636506989,
+      0.0506801187398187,
+      0.0506801187398187,
+      0.0506801187398187,
+      0.0506801187398187
+    ],
+    "bmi": [
+      -0.00620595413580824,
+      0.0369065288194278,
+      -0.00405032998804645,
+      0.0519958978537604,
+      -0.02021751109626,
+      -0.0245287593917836,
+      0.17055522598066,
+      0.045529025410475,
+      -0.0902752958985185,
+      0.0153502873418098
+    ],
+    "bp": [
+      -0.015999222636143,
+      0.0218723549949558,
+      -0.0125563519424068,
+      0.0792535333386559,
+      -0.00222773986119799,
+      -0.0263278347173518,
+      0.0149866136074833,
+      0.0218723549949558,
+      -0.0573136709609782,
+      -0.0194420933298793
+    ],
+    "s1": [
+      0.125018703134293,
+      -0.0249601584096305,
+      0.103003457403075,
+      0.054845107366035,
+      0.0383336730676214,
+      0.0988755988284711,
+      0.0300779559184146,
+      0.10988322169408,
+      -0.0249601584096305,
+      0.0369577202094203
+    ],
+    "s2": [
+      0.125198101136752,
+      -0.0166581520539057,
+      0.0487898764601065,
+      0.0365770864503148,
+      0.05317395492516,
+      0.0941964034195887,
+      0.033758750294209,
+      0.0888728795691667,
+      -0.0304366843726451,
+      0.0481635795365275
+    ],
+    "s3": [
+      0.0191869970174533,
+      0.000778807997017968,
+      0.056003375058324,
+      -0.0765355858888105,
+      -0.00658446761115617,
+      0.0707299262746723,
+      -0.0213110188275045,
+      0.000778807997017968,
+      -0.00658446761115617,
+      0.0191869970174533
+    ],
+    "s4": [
+      0.0343088588777263,
+      -0.0394933828740919,
+      -0.00259226199818282,
+      0.141322109417863,
+      0.0343088588777263,
+      -0.00259226199818282,
+      0.0343088588777263,
+      0.0343088588777263,
+      -0.00259226199818282,
+      -0.00259226199818282
+    ],
+    "s5": [
+      0.0324332257796019,
+      -0.0225121719296605,
+      0.0844952822124031,
+      0.098646374304928,
+      -0.00514530798026311,
+      -0.02139368094036,
+      0.0336568129023847,
+      0.0741925366900307,
+      0.024052583226893,
+      -0.0307512098645563
+    ],
+    "s6": [
+      -0.0052198044153011,
+      -0.0217882320746399,
+      -0.0176461251598052,
+      0.0610539062220542,
+      -0.0093619113301358,
+      0.00720651632920303,
+      0.0320591578182113,
+      0.0610539062220542,
+      0.00306440941436832,
+      -0.00107769750046639
+    ]
+  }
+}
