+++ docker_images/sklearn/tests/generators/samples/textclassification-hist_gradient_boosting-1.0.2-output.json
@@ -1,0 +1,84 @@
+[
+  [
+    {
+      "label": "0",
+      "score": 0.007188341023581015
+    },
+    {
+      "label": "1",
+      "score": 0.03422080165466977
+    },
+    {
+      "label": "2",
+      "score": 0.04593990453288178
+    },
+    {
+      "label": "3",
+      "score": 0.02403056643079793
+    },
+    {
+      "label": "4",
+      "score": 0.03209593726948873
+    },
+    {
+      "label": "5",
+      "score": 0.1118272840896114
+    },
+    {
+      "label": "6",
+      "score": 0.445567936909444
+    },
+    {
+      "label": "7",
+      "score": 0.02455902207117283
+    },
+    {
+      "label": "8",
+      "score": 0.011669934009514884
+    },
+    {
+      "label": "9",
+      "score": 0.030182113638619834
+    },
+    {
+      "label": "10",
+      "score": 0.0248637811927033
+    },
+    {
+      "label": "11",
+      "score": 0.04262809835525814
+    },
+    {
+      "label": "12",
+      "score": 0.0524577232869705
+    },
+    {
+      "label": "13",
+      "score": 0.03816834348974906
+    },
+    {
+      "label": "14",
+      "score": 0.019925741393428644
+    },
+    {
+      "label": "15",
+      "score": 0.009631384752144945
+    },
+    {
+      "label": "16",
+      "score": 0.014821126836993737
+    },
+    {
+      "label": "17",
+      "score": 0.012278330051148537
+    },
+    {
+      "label": "18",
+      "score": 0.011661211580574562
+    },
+    {
+      "label": "19",
+      "score": 0.006282417431246468
+    }
+  ]
+]
