+++ docker_images/sklearn/tests/generators/samples/textclassification-hist_gradient_boosting_classifier-latest-output.json
@@ -1,0 +1,12 @@
+[
+  6,
+  14,
+  14,
+  17,
+  17,
+  4,
+  14,
+  4,
+  13,
+  8
+]
