+++ docker_images/sklearn/tests/generators/samples/textclassification-logistic_regression-1.0.2-output.json
@@ -1,0 +1,84 @@
+[
+  [
+    {
+      "label": "0",
+      "score": 0.011504069644515289
+    },
+    {
+      "label": "1",
+      "score": 0.31411373675930093
+    },
+    {
+      "label": "2",
+      "score": 0.04701296328107312
+    },
+    {
+      "label": "3",
+      "score": 0.04859119356279581
+    },
+    {
+      "label": "4",
+      "score": 0.04670644020646418
+    },
+    {
+      "label": "5",
+      "score": 0.04975572804120902
+    },
+    {
+      "label": "6",
+      "score": 0.1320171147998885
+    },
+    {
+      "label": "7",
+      "score": 0.0691224338942939
+    },
+    {
+      "label": "8",
+      "score": 0.018680606264140507
+    },
+    {
+      "label": "9",
+      "score": 0.025537288354315006
+    },
+    {
+      "label": "10",
+      "score": 0.009272227756670208
+    },
+    {
+      "label": "11",
+      "score": 0.029191796517024207
+    },
+    {
+      "label": "12",
+      "score": 0.03303225898006273
+    },
+    {
+      "label": "13",
+      "score": 0.0812628690602073
+    },
+    {
+      "label": "14",
+      "score": 0.014160011994658371
+    },
+    {
+      "label": "15",
+      "score": 0.04754806411695202
+    },
+    {
+      "label": "16",
+      "score": 0.010401209475140446
+    },
+    {
+      "label": "17",
+      "score": 0.002283619697068288
+    },
+    {
+      "label": "18",
+      "score": 0.008539482237043945
+    },
+    {
+      "label": "19",
+      "score": 0.0012668853571762167
+    }
+  ]
+]
