+++ docker_images/sklearn/tests/test_api.py
@@ -130,7 +130,7 @@ TEST_CASES = {
             "has_config": False,
             "old_sklearn": True,
             "accepts_nan": False,
-            "loads_on_new_sklearn": True,
+            "loads_on_new_sklearn": False,
         },
         "skops-tests/tabularregression-sklearn-1.0-linear_regression-with-config": {
             "input": "tabularregression-1.0-input.json",
@@ -197,7 +197,7 @@ TEST_CASES = {
         },
         "skops-tests/textclassification-sklearn-1.0-hist_gradient_boosting-with-config": {
             "input": "textclassification-1.0-input.json",
-            "output": "textclassification-hist_gradient_boosting-1.0-output.json",
+            "output": "textclasssification-hist_gradient_boosting-1.0-output.json",
             "has_config": True,
             "old_sklearn": True,
             "loads_on_new_sklearn": False,
